+++ app/src/main/java/com/blucor/tcthecontractor/project/workorderbilling/fragments/WorkOrderFragment.java
@@ -273,9 +273,9 @@ public class WorkOrderFragment extends Fragment {
             @Override
             public void onTextChanged(CharSequence s, int start, int before, int count) {
                 try {
-                    float rate = Float.parseFloat("" + s);
-                    float qty = Float.parseFloat(et_qty.getText().toString());
-                    float amt = rate * qty;
+                    long rate = Long.parseLong("" + s);
+                    long qty = Long.parseLong(et_qty.getText().toString());
+                    long amt = rate * qty;
                     et_amount.setText("" + amt);
                 }catch (Exception exception) {
                     Log.e("TextWatcher",""+exception.getMessage());
@@ -297,9 +297,9 @@ public class WorkOrderFragment extends Fragment {
             @Override
             public void onTextChanged(CharSequence s, int start, int before, int count) {
                 try {
-                    float qty = Float.parseFloat("" + s);
-                    float rate = Float.parseFloat(et_rate.getText().toString());
-                    float amt = rate * qty;
+                    long qty =Long.parseLong("" + s);
+                    long rate = Long.parseLong(et_rate.getText().toString());
+                    long amt = rate * qty;
                     et_amount.setText("" + amt);
                 }catch (Exception exception) {
                     Log.e("TextWatcher",""+exception.getMessage());
@@ -343,7 +343,7 @@ public class WorkOrderFragment extends Fragment {
             footer_view = getFooterViewForTotalAmount();
             rv_work_order.addHeaderView(footer_view);
         } else {
-            float tot_amount = getTotalAmount();
+            long tot_amount = getTotalAmount();
             tv_footer_total.setText(""+tot_amount);
         }
     }
@@ -351,13 +351,13 @@ public class WorkOrderFragment extends Fragment {
     private View getFooterViewForTotalAmount() {
         View footer_view = LayoutInflater.from(mActivity).inflate(R.layout.work_order_list_total_item,null);
         tv_footer_total = footer_view.findViewById(R.id.tv_work_order_list_total_item);
-        float tot_amount = getTotalAmount();
+        long tot_amount = getTotalAmount();
         tv_footer_total.setText(""+tot_amount);
         return footer_view;
     }
 
-    private float getTotalAmount() {
-        float tot_amount = 0;
+    private long getTotalAmount() {
+        long tot_amount = 0;
         for (int i = 0; i < workOrders.size(); i++) {
             WorkOrderModel model = (WorkOrderModel) workOrders.get(i);
             tot_amount = tot_amount + model.amount;
@@ -388,18 +388,18 @@ public class WorkOrderFragment extends Fragment {
 
         if (is_edit) {
             WorkOrderModel edit_work_order = workOrders.get(edit_position);
-            edit_work_order.setAmount(Float.parseFloat(et_amount.getText().toString()));
-            edit_work_order.setQuantity(Integer.parseInt(et_qty.getText().toString()));
-            edit_work_order.setRate(Float.parseFloat(et_rate.getText().toString()));
+            edit_work_order.setAmount(Long.parseLong(et_amount.getText().toString()));
+            edit_work_order.setQuantity(Long.parseLong(et_qty.getText().toString()));
+            edit_work_order.setRate(Long.parseLong(et_rate.getText().toString()));
             edit_work_order.setUnit_id(unit_id);
             edit_work_order.setUnit(et_unit.getText().toString());
             edit_work_order.setWork_description(et_workdesc.getText().toString());
             editWorkOrder(edit_work_order);
         } else {
             WorkOrderModel workOrder = new WorkOrderModel();
-            workOrder.setAmount(Float.parseFloat(et_amount.getText().toString()));
-            workOrder.setQuantity(Integer.parseInt(et_qty.getText().toString()));
-            workOrder.setRate(Float.parseFloat(et_rate.getText().toString()));
+            workOrder.setAmount(Long.parseLong(et_amount.getText().toString()));
+            workOrder.setQuantity(Long.parseLong(et_qty.getText().toString()));
+            workOrder.setRate(Long.parseLong(et_rate.getText().toString()));
             workOrder.setUnit_id(unit_id);
             workOrder.setUnit(et_unit.getText().toString());
             workOrder.setWork_description(et_workdesc.getText().toString());
